Set up the database seeder to load CSV data into a Pokemon table.
We're now using Sqlite to store the Pokemon data from the CSV file and use EntityFramework as the ORM to manipulate and expose the data that we have stored into the RDBMS.

Right now, we're using Sqlite as an in-memory database. We don't have any need, as it turns out, to have a specific connection string. It'll stay on RAM and be pretty fast (there won't be any persistence but the file is small and importing the data is quite fast - especially since we're focused on building an MVP).